+++ includes/classes/Setup.php
@@ -36,6 +36,7 @@ class Setup
   require_once 'awm-list-tables/class-list-table.php';
   require_once 'awm-customizer/class-customizer.php';
   require_once 'ewp-third-party/class-wpml.php';
+  require_once 'ewp-third-party/class-wp-rocket.php';
   require_once 'ewp-gutenburg/class-register.php';
   require_once 'wp-cli/class-cli-commands.php';
   require_once 'dev-tools/init.php';